Extract email subscription as component
