+++ src/components/EmailSubscription/index.module.scss
@@ -1,0 +1,28 @@
+.paperPlaneButton {
+  background-color: transparent;
+  border: none;
+  position: absolute;
+  right: 0;
+  padding-right: 1em;
+  height: 100%; // tips from Sheldon, make the height same as input! :)
+  z-index: 4;
+}
+
+.emailTextfieldWithRightButton {
+  display: inline-block;
+  border-top-right-radius: 8px !important;
+  border-bottom-right-radius: 8px !important;
+  border-bottom-left-radius: 8px !important;
+  border-top-left-radius: 8px !important;
+  padding-right: 56px; // prevent input text from going under the button
+}
+
+.paperPlaneSvg {
+  display: inline-block;
+  background: url('/img/paper-plane.svg') no-repeat top left;
+  background-size: contain;
+  cursor: pointer;
+  height: 24px;
+  width: 24px;
+  vertical-align: middle;
+}
