+++ src/pages/index.module.scss
@@ -15,45 +15,3 @@
   height: 10rem;
   object-fit: cover;
 }
-
-.paperPlaneButton {
-  background-color: transparent;
-  border: none;
-  position: absolute;
-  right: 0;
-  padding-right: 1em;
-  height: 100%; // tips from Sheldon, make the height same as input! :)
-  z-index: 4;
-}
-
-.emailTextfieldWithRightButton {
-  display: inline-block;
-  border-top-right-radius: 8px !important;
-  border-bottom-right-radius: 8px !important;
-  border-bottom-left-radius: 8px !important;
-  border-top-left-radius: 8px !important;
-  padding-right: 56px; // prevent input text from going under the button
-}
-
-.addRightView {
-  display: flex;
-  border: 1px solid #ced4da;
-  border-left: none;
-  border-top-right-radius: 8px !important;
-  border-bottom-right-radius: 8px !important;
-  
-}
-
-.noRightBorder {
-  border-right: none;
-}
-
-.paperPlaneSvg {
-  display: inline-block;
-  background: url('/img/paper-plane.svg') no-repeat top left;
-  background-size: contain;
-  cursor: pointer;
-  height: 24px;
-  width: 24px;
-  vertical-align: middle;
-}
